+++ tests/Unit/Delivery/QueryTest.php
@@ -13,19 +13,6 @@ class QueryTest extends \PHPUnit_Framework_TestCase
     /**
      * @covers Contentful\Delivery\Query::__construct
      * @covers Contentful\Delivery\Query::setInclude
-     * @covers Contentful\Delivery\Query::getInclude
-     */
-    public function testGetSetInclude()
-    {
-        $query = new Query;
-        $query->setInclude(50);
-
-        $this->assertEquals(50, $query->getInclude());
-    }
-
-    /**
-     * @covers Contentful\Delivery\Query::__construct
-     * @covers Contentful\Delivery\Query::setInclude
      * @covers Contentful\Delivery\Query::getQueryData
      * @covers Contentful\Query::getQueryData
      * @covers Contentful\Query::getQueryString
@@ -36,19 +23,6 @@ class QueryTest extends \PHPUnit_Framework_TestCase
         $query->setInclude(50);
 
         $this->assertEquals('include=50', $query->getQueryString());
-    }
-
-    /**
-     * @covers Contentful\Delivery\Query::__construct
-     * @covers Contentful\Delivery\Query::setLocale
-     * @covers Contentful\Delivery\Query::getLocale
-     */
-    public function testGetSetLocale()
-    {
-        $query = new Query;
-        $query->setLocale('de-DE');
-
-        $this->assertEquals('de-DE', $query->getLocale());
     }
 
     /**