+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased](https://github.com/contentful/ContentfulBundle/compare/0.7.0-beta...HEAD)
 
+### Removed
+* Removed all get* methods except `getQueryData()` and `getQueryString()` from the various query classes. **[BREAKING]**
+
 ## [0.7.0-beta](https://github.com/contentful/contentful.php/tree/0.7.0-beta) (2017-04-06)
 
 **ATTENTION**: This release contains breaking changes. Please take extra care when updating to this version.